Extract subtask execution scheduler
